modules/accessModule/requirements.txt
azure-iot-device~=2.7.0
smbus2==0.2.1
python-dotenv
RPi.GPIO
